fix: resolve entity-DB mismatches and middleware gaps causing test failures - Map Invoice entity to actual DB columns (SalespersonPersonID, PackedByPersonID) instead of non-existent DeliveredByPersonID - Change TotalDryItems/TotalChillerItems from decimal to int to match DB - Handle SqlException with error number 0 and nested SqlExceptions in ExceptionHandlingMiddleware for proper 503 responses - Fix HealthController test to override config instead of services - Fix PaymentController test to use dynamically fetched valid ID

+++ backend/WideWorldImporters.Api/Models/Dtos/DeliveryDtos.cs
@@ -19,8 +19,8 @@ namespace WideWorldImporters.Api.Models.Dtos
         public string CustomerName { get; set; }
         public string DriverName { get; set; }
         public DateTime InvoiceDate { get; set; }
-        public decimal TotalDryItems { get; set; }
-        public decimal TotalChillerItems { get; set; }
+        public int TotalDryItems { get; set; }
+        public int TotalChillerItems { get; set; }
         public List<InvoiceLineDto> Lines { get; set; }
     }
 }
